Factor output status reporting into separate component
