+++ src/Microsoft.DotNet.ImageBuilder/src/Models/McrStatus/ImageStatus.cs
@@ -5,17 +5,18 @@
 using System;
 using Newtonsoft.Json;
 
+#nullable enable
 namespace Microsoft.DotNet.ImageBuilder.Models.McrStatus
 {
     public class ImageStatus
     {
         [JsonProperty(PropertyName = "Id")]
-        public string OnboardingRequestId { get; set; }
-        public string SourceRepository { get; set; }
-        public string TargetRepository { get; set; }
+        public string OnboardingRequestId { get; set; } = string.Empty;
+        public string SourceRepository { get; set; } = string.Empty;
+        public string TargetRepository { get; set; } = string.Empty;
         public DateTime QueueTime { get; set; }
-        public string Tag { get; set; }
+        public string? Tag { get; set; }
         public StageStatus OverallStatus { get; set; }
-        public string FailureReason { get; set; }
+        public string? FailureReason { get; set; }
     }
 }
